python benchmark tuning
* add initial workflow npbench

* reset base path

* update build dependencies

* extend error

* use docc ci regions

* update harness + add benchmark script

* automate summary

* deactivate workflow

* add heuristic

* dot expansion

* update test parsing

* skip nussinov

* gemm expansion

* add new benchmarks

* use standard c types

* use 64bit int

* 32 to64 bit int

* update summary

* adapt unit tests

* fix dot parsing

* format

* fix docc regions

* env setup

* update counts

* spmv counter

+++ python/benchmarks/npbench/polybench/test_floyd_warshall.py
@@ -5,7 +5,7 @@ from benchmarks.npbench.harness import SDFGVerification, run_benchmark, run_pyte
 PARAMETERS = {"S": {"N": 200}, "M": {"N": 400}, "L": {"N": 850}, "paper": {"N": 2800}}
 
 
-def initialize(N, datatype=np.int32):
+def initialize(N, datatype=np.int64):  # originally np.int32
     path = np.fromfunction(lambda i, j: i * j % 7 + 1, (N, N), dtype=datatype)
     for i in range(N):
         for j in range(N):
